requirements.txt
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# Database — need 2.0+ for DeclarativeBase
sqlalchemy>=2.0.0

# Vector Store (lightweight, no torch needed)
faiss-cpu>=1.7.4
numpy>=1.24.0

# Data Processing
pandas>=2.1.0
python-dateutil>=2.8.0

# LLM + Embeddings (Google Gemini — new SDK v1.0+)
google-genai>=1.0.0

# Frontend
streamlit>=1.30.0

# File handling
python-multipart>=0.0.9
aiofiles>=23.2.0

# HTTP
httpx>=0.27.0
requests>=2.31.0
